Move SW files into sdk and add fetch listener
